Improve readium shared library
